Added some animations, complicated the movement, fixed some issues with collision
Animations added are: Idle, Walk, Jump and Fall. Currently functional, however, there's bugs regarding IsGrounded like saying the player isn't when they're still colliding with the ground. Simple fix though, will get around to it next class.

The movement added was the ability to jump! That's about it really, but the jump and falling animations seem to be working okay for the most part.

When it comes to map fixes, I just fixed the ability to jump through some platforms by enabling "UsedByEffector". That's all!