Coin added, fixed fireball and added animations
Specifically, I changed a majority of the C# files to support all the different things I added. I also added an audio system to the coin that plays when you touch it. Also added a cool animation to the coin when you pick it up. It could definitely be done better once I learn how to do coroutines.